scrum-2764 dump prod

+++ agr_literature_service/lit_processing/data_export/dump_database.py
@@ -1,20 +1,15 @@
 import argparse
 import logging
-import json
-from os import environ, makedirs, path, rename, remove
+from os import environ
 from dotenv import load_dotenv
-from datetime import datetime, date
-from agr_literature_service.lit_processing.utils.s3_utils import upload_file_to_s3, download_file_from_s3, delete_file_from_s3, file_exist_from_s3
-from datetime import datetime, timedelta
+from agr_literature_service.lit_processing.utils.s3_utils import upload_file_to_s3, delete_file_from_s3, file_exist_from_s3
+from datetime import timedelta, date
 import os
-# init_tmp_dir()
 
 logging.basicConfig(format='%(message)s')
 log = logging.getLogger(__name__)
 log.setLevel(logging.INFO)
-
 load_dotenv()
-
 s3_bucket = 'agr-literature'
 sub_bucket = 'prod/database_dump/'
 latest_bucket = sub_bucket + 'latest/'
@@ -22,21 +17,30 @@ lastweek_bucket = sub_bucket + 'last_week/'
 monthly_bucket = sub_bucket + 'monthly_archive/'
 ondemand_bucket = sub_bucket + 'ondemand/'
 
-def dump_database(dump_type ="ondemand"):  # noqa: C901
+
+def dump_database(dump_type="ondemand"):  # noqa: C901
+    env_state = environ.get('ENV_STATE', "")
+    if dump_type == 'cron' and env_state != 'prod':
+        log.info("dump for production database only")
+        return False
     database = environ.get('PSQL_DATABASE', "")
     host = environ.get('PSQL_HOST', "")
     username = environ.get('PSQL_USERNAME', "")
     password = environ.get('PSQL_PASSWORD', "")
     port = environ.get('PSQL_PORT', "")
-    # file_name_date_7 = database + "_" + str(date.today() - timedelta(days=7)) + "_.sql"
-    file_name_date_7 = database + "_" + str(date.today()) + "_.sql"
-    file_name = database + "_" + str(date.today()) + "_.sql"
+    file_name_date_7 = database + "_" + str(date.today() - timedelta(days=7)) + ".sql"
+    file_name = database + "_" + str(date.today()) + ".sql"
     cmd = "PGPASSWORD=" + password + " pg_dump -Fc --clean -h " + host + " -p " + port + " -U " + username + " " + database + "  > " + file_name
     log.info(cmd)
     os.system(cmd)
     if dump_type == 'cron':
         s3_filename = lastweek_bucket + file_name
         upload_file_to_s3(file_name, s3_bucket, s3_filename)
+        ## upload file to monthly bucket if it is first day of the month
+        todayDate = date.today()
+        if todayDate.day == 1:
+            s3_filename_monthly = monthly_bucket + file_name
+            upload_file_to_s3(file_name, s3_bucket, s3_filename_monthly, 'GLACIER_IR')
         # delete local file after upload to s3
         try:
             os.remove(file_name)
@@ -44,21 +48,12 @@ def dump_database(dump_type ="ondemand"):  # noqa: C901
             log.info("fail to delete local file")
             return False
 
-        # put the file older than 7 days into monthly if exists
-        file_date_7_exist = file_exist_from_s3(file_name_date_7)
+        ## delete day7 file from lastweek folder
+        s3_filename_day7 = lastweek_bucket + file_name_date_7
+        file_date_7_exist = file_exist_from_s3(s3_bucket, s3_filename_day7)
         if file_date_7_exist:
-            s3_filename_day7 = lastweek_bucket + file_name_date_7
-            file_name_downloaded = database + "_" + str(date.today()) + "_downloaded_.sql"
-            log.info(s3_filename_day7)
-            downloaded = download_file_from_s3(file_name_downloaded, s3_bucket, s3_filename_day7)
-            if downloaded:
-                s3_filename_monthly = monthly_bucket + file_name_downloaded
-                upload_file_to_s3(file_name_downloaded, s3_bucket, s3_filename_monthly, 'DEEP_ARCHIVE')
-                delete_file_from_s3(s3_bucket, s3_filename_day7)
-                try:
-                    os.remove(file_name_downloaded)
-                except OSError:
-                    return False
+            delete_file_from_s3(s3_bucket, s3_filename_day7)
+
     return True
